Refactor: liblrmd: separate events API into own header file
This will help avoid circular includes
