+++ include/crm/Makefile.am
@@ -1,5 +1,5 @@
 #
-# Copyright 2004-2021 the Pacemaker project contributors
+# Copyright 2004-2023 the Pacemaker project contributors
 #
 # The version control history for this file may have further details.
 #
@@ -11,11 +11,18 @@ MAINTAINERCLEANFILES	= Makefile.in
 
 headerdir=$(pkgincludedir)/crm
 
-header_HEADERS		= cib.h cluster.h compatibility.h crm.h \
-			  lrmd.h msg_xml.h services.h stonith-ng.h \
+header_HEADERS		= cib.h			\
+			  cluster.h		\
+			  compatibility.h	\
+			  crm.h			\
 			  crm_compat.h		\
+			  lrmd.h		\
+			  lrmd_events.h		\
+			  msg_xml.h		\
 			  msg_xml_compat.h	\
-			  services_compat.h
+			  services.h		\
+			  services_compat.h	\
+			  stonith-ng.h
 
 noinst_HEADERS		= lrmd_internal.h services_internal.h
